# docs/swagger.yaml
basePath: /api/
definitions:
  apierr.ApiErr:
    properties:
      code:
        type: integer
      message: {}
    type: object
  dto.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 12
        type: string
      old_password:
        minLength: 12
        type: string
    required:
    - new_password
    - old_password
    type: object
  dto.DeleteUserRequest:
    properties:
      password:
        minLength: 12
        type: string
    required:
    - password
    type: object
  dto.IdResponse:
    properties:
      id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      message:
        example: readme created successfully
        type: string
    required:
    - id
    type: object
...
  dto.SearchWidgetRequestDoc:
    properties:
      amount:
        minimum: 1
        type: integer
      filter:
        $ref: '#/definitions/dto.filterWidgetsFields'
      page:
        minimum: 1
        type: integer
      query:
        type: string
      sort:
        $ref: '#/definitions/dto.sortWidgetsFields'
    required:
    - amount
    - page
    type: object
...
  dto.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: success
        type: string
    type: object
  dto.TemplateInfo:
    properties:
      description:
        minLength: 1
        type: string
      id:
        type: string
      image:
        type: string
      is_public:
        type: boolean
      last_update_time:
        type: string
      likes:
        minimum: 0
        type: integer
      num_of_users:
        minimum: 0
        type: integer
      title:
        minLength: 1
        type: string
    required:
    - description
    - id
    - image
    - is_public
    - last_update_time
    - likes
    - num_of_users
    - title
    type: object
...
  dto.UserResponse:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: string
      nickname:
        minLength: 1
        type: string
      num_of_readmes:
        minimum: 0
        type: integer
      num_of_templates:
        minimum: 0
        type: integer
      templates:
        items:
          $ref: '#/definitions/dto.TemplateInfo'
        minItems: 0
        type: array
      time_of_register:
        type: string
    required:
    - avatar
    - email
    - id
    - nickname
    - num_of_readmes
    - num_of_templates
    - templates
    - time_of_register
    type: object
...
  dto.WidgetResponse:
    properties:
      description:
        type: string
      id:
        type: string
      image:
        type: string
      likes:
        minimum: 0
        type: integer
      num_of_users:
        minimum: 0
        type: integer
      title:
        type: string
    required:
    - description
    - id
    - image
    - likes
    - num_of_users
    - title
    type: object
...
  dto.filterWidgetsFields:
    properties:
      Tags:
        items:
          type: string
        type: array
      Types:
        items:
          type: string
        type: array
    type: object
...
  dto.sortWidgetsFields:
    properties:
      Likes:
        example: desc/asc
        type: string
      NumOfUsers:
        example: desc/asc
        type: string
    type: object
...
  models.Widget:
    properties:
      description:
        type: string
      id:
        type: string
      image:
        type: string
      likes:
        type: integer
      link:
        type: string
      num_of_users:
        type: integer
      tags:
        additionalProperties: {}
        type: object
      title:
        type: string
      type:
        type: string
    type: object
host: localhost:3333
info:
  contact:
    email: readmeow@gmail.com
    name: READMEOW Support
  description: Documentation to READMEOW API
  title: REODMEOW API
  version: 1.0.0
paths:
...
  /api/users:
    delete:
      consumes:
      - application/json
      description: Delete current user account
      parameters:
      - description: Delete user request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Delete User
      tags:
      - Users
    patch:
      consumes:
      - multipart/form-data
      description: Update user's profile (nickname, avatar, etc.)
      parameters:
      - format: binary
        in: formData
        name: avatar
        type: string
      - in: formData
        name: id
        required: true
        type: string
      - in: formData
        maxLength: 80
        minLength: 1
        name: nickname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - Users
  /api/users/{user}:
    get:
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Get User
      tags:
      - Users
  /api/users/password:
    patch:
      consumes:
      - application/json
      description: Change password for current user
      parameters:
      - description: Change password request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Change User Password
      tags:
      - Users
  /api/widgets:
    get:
      consumes:
      - application/json
      description: Searching widgets
      parameters:
      - description: Search widgets request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SearchWidgetRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: List of widgets
          schema:
            items:
              $ref: '#/definitions/dto.WidgetResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      summary: Search Widgets
      tags:
      - Widgets
    post:
      consumes:
      - multipart/form-data
      description: Creating a new widget
      parameters:
      - in: formData
        maxLength: 1000
        minLength: 1
        name: description
        required: true
        type: string
      - format: binary
        in: formData
        name: image
        required: true
        type: string
      - in: formData
        name: link
        required: true
        type: string
      - in: formData
        name: tags
        required: true
        type: string
      - in: formData
        maxLength: 80
        minLength: 1
        name: title
        required: true
        type: string
      - in: formData
        maxLength: 50
        minLength: 1
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.IdResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helpers.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helpers.ApiErr'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/helpers.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/helpers.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/helpers.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Create Widget
      tags:
      - Widgets
  /api/widgets/{widget}:
    get:
      description: Get widget by ID
      parameters:
      - description: Widget ID
        in: path
        name: widget
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Widget data
          schema:
            $ref: '#/definitions/models.Widget'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Get Widget
      tags:
      - Widgets
  /api/widgets/dislike/{widget}:
    patch:
      description: Dislike widget by ID
      parameters:
      - description: Widget ID
        in: path
        name: widget
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Dislike Widget
      tags:
      - Widgets
  /api/widgets/favorite:
    get:
      consumes:
      - application/json
      description: Fetch favorite widgets of current user
      parameters:
      - in: query
        minimum: 1
        name: amount
        required: true
        type: integer
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of favorite widgets
          schema:
            items:
              $ref: '#/definitions/dto.WidgetResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Fetch Favorite Widgets
      tags:
      - Widgets
  /api/widgets/like/{widget}:
    patch:
      description: Like widget by ID
      parameters:
      - description: Widget ID
        in: path
        name: widget
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Like Widget
      tags:
      - Widgets
swagger: "2.0"